codesniff/backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Code Analysis
tree-sitter==0.20.4
tree-sitter-python==0.20.4

# ML & Embeddings
torch==2.1.0
transformers==4.35.0
sentence-transformers==2.2.2

# Vector Search
faiss-cpu==1.7.4
numpy==1.24.3

# Database
aiosqlite==0.19.0

# Utilities
pydantic==2.5.0
python-dotenv==1.0.0
loguru==0.7.2
tqdm==4.66.1

# AI Chat
groq==0.4.1

# Development
pytest==7.4.3
httpx==0.25.1
